Create Sign in, sign out and sign up calls. - Rename file firebase-init.ts to firebase-auth and move to firebase subfolder.
