--- src/app/services/firebase-init.ts
@@ -1,14 +1,0 @@
-﻿import { initializeApp } from "firebase/app";
-
-const firebaseConfig = {
-    apiKey: process.env.REACT_APP_APIKEY,
-    authDomain: process.env.REACT_APP_AUTHDOMAIN,
-    projectId: process.env.REACT_APP_PROJECTID,
-    storageBucket: process.env.REACT_APP_STORAGEBUCKET,
-    messagingSenderId: process.env.REACT_APP_SENDERID,
-    appId: process.env.REACT_APP_APPID,
-    measurementId: process.env.REACT_APP_ANALYTICS
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
