+++ src/app/services/firebase/firebase-auth.ts
@@ -1,0 +1,53 @@
+﻿import { initializeApp } from "firebase/app";
+import {createUserWithEmailAndPassword, getAuth, signInWithEmailAndPassword} from "firebase/auth";
+
+const firebaseConfig = {
+    apiKey: process.env.REACT_APP_APIKEY,
+    authDomain: process.env.REACT_APP_AUTHDOMAIN,
+    projectId: process.env.REACT_APP_PROJECTID,
+    storageBucket: process.env.REACT_APP_STORAGEBUCKET,
+    messagingSenderId: process.env.REACT_APP_SENDERID,
+    appId: process.env.REACT_APP_APPID,
+    measurementId: process.env.REACT_APP_ANALYTICS
+};
+
+// Initialize Firebase
+const app = initializeApp(firebaseConfig);
+//Auth reference for Firebase
+export const auth = getAuth(app);
+
+/**
+ * Login to Firebase Auth with the user's email and password.
+ * @param email {string} - The user's account email.
+ * @param password {string} - The user's account password.
+ */
+export const logInWithEmailAndPassword = (email: string, password: string) => {
+    signInWithEmailAndPassword(auth, email, password)
+        .then((userCredential) => {
+            // Signed in 
+            const user = userCredential.user;
+        })
+        .catch((error) => {
+            const errorCode = error.code;
+            const errorMessage = error.message;
+        });
+}
+
+export const logOut = () => auth.signOut();
+
+/**
+ * Sign up via Firebase Auth with the user's email and password.
+ * @param email {string} - The new user's account email.
+ * @param password {string} - The new user's account password.
+ */
+export const signUpWithEmailAndPassword = (email: string, password: string) => {
+    createUserWithEmailAndPassword(auth, email, password)
+        .then((userCredential) => {
+            // Signed up 
+            const user = userCredential.user;
+        })
+        .catch((error) => {
+            const errorCode = error.code;
+            const errorMessage = error.message;
+        });
+}
